Add a new fallback for user agents that only provide a fully qualified language tag, and not the general language family (which is what we are likely to have translated for).
Only the first four steps are 'supported', we are just guessing 
at this point, so we can freely take this reasonable assumption.

If the UA had provided a well-filled list of acceptable language, 
we shouldn't reach to this point.

This was prompted by the data gathering related to this mail:
 http://lists.wikimedia.org/pipermail/wikilovesmonuments/2012-October/005256.html



git-svn-id: https://svn.wikimedia.org/svnroot/mediawiki/trunk/tools/ToolserverI18N@115775 dd0e9695-b195-4be7-bd10-2dea1a65a6b6