Update en.json
Added missing Commonscat entry and changing the source-monuments-list entry since lists do not need to live on Wikipedia since https://gerrit.wikimedia.org/r/#/c/286156/

"source-monuments-list-on-wikipedia" is no longer used but it is unclear to me whether it is better to leave it (so that the text can be used as a suggestion for the new similar message) or if it should be removed. It is also unclear in which order this and XXX should be merged/deployed.