js-env: Sort domain list to increase cache opportunities
Since 213cbba we have a limited time window to allow for batch-requesting
multiple domains in one request. However these will come from unrelated
scripts which may not always load in the same order on every page.

Sort the list to improve chances of a cache hit of the same url.

This way if a user loads two gadgets on all pages, even if they don't
always execute in the same order, as long as they load within a given
time period they'll always end up in the same batch request and with
the same url.