Intuition: Fix file_get_contents warning in loadMessageFile
This often happens for domains that are not yet fully translated.
When setting the user preference centrally to a language that doesn't
yet have any translations for a particular domain, loading that domain
in a tool will fail due to the specific JSON file not existing
in that tool's i18n directory.

Fixes #81