Fix string checking bug and refreshLang param
The initLangSelect() method was treating the return value of
strstr() as an integer, which was resulting in a warning:
"A non-numeric value encountered at Intuition.php:1347"

This patch fixes this and makes it clearer what's happening.
It also fixes refreshLang(), which was previously not working.

A phpcs exception is also removed.