Add {{BIDI}} function, similar to what exists in MW language
This allows for wrapping parameters in {{BIDI:$1}} so they are
wrapped automatically with control characters, either RTL or LTR.

The code auto identifies the given string as RTL or LTR strong
characters, and adds control characters accordingly.

** CURRENTLY THERE IS AN ISSUE WITH REGEXP **
The regular expression used to discover the function inside {{}}
is assuming that all functions have parameters with pipes, so
creating {{BIDI:foo}} doesn't seem to work properly. Changing
the regexp creates a problem with the grouping of $matches
that the code relies on.

I could use help with this.