Intuition: setUseRequestParam now always sets the value
If given an invalid value, it is interpreted as false by default.

The return value is removed because, aside from tests, it was never
used, and I don't want code to have to check this. If we want error
handling, we should use exceptions. E.g. by setting argument type
as 'bool', ref https://github.com/Krinkle/intuition/issues/95.