Intuition: Refocus getAvailableLangs() on local domains
This is a left-over from when most message domains shipped with
Intuition. The generating of the langlist (a subset of Names.php)
is becoming a burdon to maintain.

In addition to requiring maintenance, it's actual value is rather
small because aside from a few legacy tools, most tools using it
would be getting an arbirary subset of languages based on message
domains not related to them.

There is actually one internal call to this method, in initLangSelect().
That call is in my opinion a bug. It was filtering values from
the user's browser preferences (exposed as Accept-Language header)
based on the "available languages".

The idea was that, if the user for example has "de, fr, en" as
their preference order (where de and fr are not fallbacks for each
other), and the current UI is translated in FR but not DE, then
we could offer that directly. Doing it based on all known languages
would pick DE which then falls back to EN directly.

This logic is fine, but should be based on the current domain
instead.