Extend lang normalization to only accept alphanums and hyphen
This prevents the language being set to an invalid value.
At the moment, it's possible to use setLang() with an invalid
language code, and therefore for getLang() to return something
invalid.

One fix might be to check the provided lang against the list of
available languages, but that might be annoying for some usages,
so instead this just makes the language code normalization more
strict: replacing anything that's not alphanumeric or hyphen
with a hyphen (and then trimming hyphens from the string). This
isn't always going to be a valid language code, but it'll always
be safe to use wherever a language code is wanted.

https://phabricator.wikimedia.org/T239871