build: Update to mediawiki-codesniffer 32.0, fix UnusedUseStatement
This might fix the PHP 8.0 build, which currently fails the following
failure with mediawiki-codesniffer 28.0.0, and only on PHP 8.0 builds,
not on any PHP 7.x build:

> /intuition/tests/phpunit/UtilTest.php
> ----------------------------------------------------------------------
>  3 | WARNING | [x] Use statement with non-compound name
>    |         |     (MediaWiki.Classes.UnusedUseStatement.NonCompoundUse)

I don't know why this doesn't happen on PHP 7.x. After updating to
mediawiki-codesniffer 31.0, it still happens, and still only on
PHP 8.0:

> /intuition/tests/phpunit/UtilTest.php
> ----------------------------------------------------------------------
>  12 | WARNING | [x] Unnecessary use statement in the same
>     |         |     namespace
>     |         |     (MediaWiki.Classes.UnusedUseStatement.UnnecessaryUse)

It seems to have been fixed in mediawiki-codesniffer 32.0.